--- src/__mocks__/@raycast/api.ts
@@ -1,14 +1,0 @@
-/**
- * Mocked functions for @raycast/api
- * Required for testing without actually using the Raycast API
- */
-export const getSelectedText = jest.fn();
-export const showToast = jest.fn();
-export const getPreferenceValues = jest.fn();
-export const Toast = {
-  Style: {
-    Animated: "Animated",
-    Success: "Success",
-    Failure: "Failure",
-  },
-};
